+++ example/lib/main.dart
@@ -1,3 +1,6 @@
+import 'dart:isolate';
+import 'dart:ui';
+
 import 'package:awesome_notifications/awesome_notifications.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
@@ -5,6 +8,7 @@ import 'package:http/http.dart' as http;
 Future<void> main() async {
   // Always initialize Awesome Notifications
   await NotificationController.initializeLocalNotifications();
+  await NotificationController.initializeIsolateReceivePort();
   runApp(const MyApp());
 }
 
@@ -42,6 +46,19 @@ class NotificationController {
         .getInitialNotificationAction(removeFromActionEvents: false);
   }
 
+  static ReceivePort? receivePort;
+  static Future<void> initializeIsolateReceivePort() async {
+    receivePort = ReceivePort('Notification action port in main isolate')
+      ..listen(
+              (silentData) => onActionReceivedImplementationMethod(silentData)
+      );
+
+    IsolateNameServer.registerPortWithName(
+        receivePort!.sendPort,
+        'notification_action_port'
+    );
+  }
+
   ///  *********************************************
   ///     NOTIFICATION EVENTS LISTENER
   ///  *********************************************
@@ -60,20 +77,41 @@ class NotificationController {
       ReceivedAction receivedAction) async {
 
     if(
-      receivedAction.actionType == ActionType.SilentAction ||
-      receivedAction.actionType == ActionType.SilentBackgroundAction
+        receivedAction.actionType == ActionType.SilentAction ||
+        receivedAction.actionType == ActionType.SilentBackgroundAction
     ){
       // For background actions, you must hold the execution until the end
       print('Message sent via notification input: "${receivedAction.buttonKeyInput}"');
       await executeLongTaskInBackground();
     }
     else {
+      // this process is only necessary when you need to redirect the user
+      // to a new page or use a valid context
+      if (receivePort == null) {
+        print('onActionReceivedMethod was called inside a parallel dart isolate.');
+        SendPort? sendPort = IsolateNameServer.lookupPortByName(
+            'notification_action_port'
+        );
+
+        if (sendPort != null) {
+          print('Redirecting the execution to main isolate process.');
+          sendPort.send(receivedAction);
+          return;
+        }
+      }
+
+      return onActionReceivedImplementationMethod(receivedAction);
+    }
+  }
+
+  static Future<void> onActionReceivedImplementationMethod(
+      ReceivedAction receivedAction
+  ) async {
       MyApp.navigatorKey.currentState?.pushNamedAndRemoveUntil(
           '/notification-page',
               (route) =>
           (route.settings.name != '/notification-page') || route.isFirst,
           arguments: receivedAction);
-    }
   }
 
   ///  *********************************************